Update pathway_errorbar.R - applying left join to properly map grouping variable to samples
A proper mapping from sample names to their corresponding groups is achieved by applying a left_join function. This requires declaring additional parameters in the pathway_errorbar function, these are: 

df_meta: The metadata data frame
meta_id_col: This is the key column for matching the abundance data with the grouping variable in the metadata. 

Function parameters df_data and meta_group_col are redundant with Group and abundance, but are useful to separate the mapping code, which starts at line 286.